+++ client/src/io/worxbend/gitea4s/api/ReleasesApi.scala
@@ -10,6 +10,8 @@ trait ReleasesApi:
 
   def release(owner: String, repo: String, id: Long): IO[GiteaError, Release]
 
+  def releaseByTag(owner: String, repo: String, tag: String): IO[GiteaError, Release]
+
   def releaseAssets(owner: String, repo: String, releaseId: Long): IO[GiteaError, Chunk[ReleaseAsset]]
 
   def releaseAsset(owner: String, repo: String, releaseId: Long, assetId: Long): IO[GiteaError, ReleaseAsset]